+++ src/main/java/me/inf32768/ultimatescaler/UltimateScaler.java
@@ -1,7 +1,8 @@
 package me.inf32768.ultimatescaler;
 
 import me.inf32768.ultimatescaler.commands.LocatePosition;
-import me.inf32768.ultimatescaler.config.WorldGenOptions;
+import me.inf32768.ultimatescaler.option.KeyBindings;
+import me.inf32768.ultimatescaler.option.UltimateScalerOptions;
 import net.fabricmc.api.ModInitializer;
 
 public class UltimateScaler implements ModInitializer {
@@ -9,7 +10,7 @@ public class UltimateScaler implements ModInitializer {
     @Override
     public void onInitialize() {
         LocatePosition.init();
-
-        WorldGenOptions.initializeConfig();
+        UltimateScalerOptions.init();
+        KeyBindings.init();
     }
 }