+++ src/main/java/me/inf32768/ultimatescaler/mixins/MixinShift.java
@@ -4,7 +4,7 @@ import net.minecraft.world.gen.densityfunction.DensityFunctionTypes;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Overwrite;
 
-import me.inf32768.ultimatescaler.config.WorldGenOptions;
+import me.inf32768.ultimatescaler.option.UltimateScalerOptions;
 
 @Mixin(DensityFunctionTypes.Shift.class)
 public abstract class MixinShift implements DensityFunctionTypes.Offset {
@@ -15,6 +15,6 @@ public abstract class MixinShift implements DensityFunctionTypes.Offset {
      */
     @Overwrite
     public double sample(NoisePos pos) {
-        return this.sample((double)pos.blockX() * WorldGenOptions.globalScale[0] + WorldGenOptions.globalOffset[0], (double)pos.blockY() * WorldGenOptions.globalScale[1] + WorldGenOptions.globalOffset[1], (double)pos.blockZ() * WorldGenOptions.globalScale[2] + WorldGenOptions.globalOffset[2]) * (double)4.0F;
+        return this.sample((double)pos.blockX() * UltimateScalerOptions.globalScale[0] + UltimateScalerOptions.globalOffset[0], (double)pos.blockY() * UltimateScalerOptions.globalScale[1] + UltimateScalerOptions.globalOffset[1], (double)pos.blockZ() * UltimateScalerOptions.globalScale[2] + UltimateScalerOptions.globalOffset[2]);
     }
 }