+++ src/main/java/me/inf32768/ultimatescaler/mixins/MixinShiftA.java
@@ -4,7 +4,7 @@ import net.minecraft.world.gen.densityfunction.DensityFunctionTypes;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Overwrite;
 
-import me.inf32768.ultimatescaler.config.WorldGenOptions;
+import me.inf32768.ultimatescaler.option.UltimateScalerOptions;
 
 @Mixin(DensityFunctionTypes.ShiftA.class)
 public abstract class MixinShiftA implements DensityFunctionTypes.Offset {
@@ -14,6 +14,6 @@ public abstract class MixinShiftA implements DensityFunctionTypes.Offset {
      */
     @Overwrite
     public double sample(NoisePos pos) {
-        return this.sample((double)pos.blockX() * WorldGenOptions.globalScale[0] + WorldGenOptions.globalOffset[0], 0.0D, (double)pos.blockZ() * WorldGenOptions.globalScale[2] + WorldGenOptions.globalOffset[2]) * 4.0D;
+        return this.sample((double)pos.blockX() * UltimateScalerOptions.globalScale[0] + UltimateScalerOptions.globalOffset[0], 0.0D, (double)pos.blockZ() * UltimateScalerOptions.globalScale[2] + UltimateScalerOptions.globalOffset[2]);
     }
 }