+++ src/main/java/me/inf32768/ultimatescaler/mixins/MixinShiftedNoiseSampler.java
@@ -8,16 +8,16 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.ModifyArgs;
 import org.spongepowered.asm.mixin.injection.invoke.arg.Args;
 
-import me.inf32768.ultimatescaler.config.WorldGenOptions;
+import me.inf32768.ultimatescaler.option.UltimateScalerOptions;
 
 @Mixin(DensityFunctionTypes.ShiftedNoise.class)
 public abstract class MixinShiftedNoiseSampler {
     @ModifyArgs(method = "sample", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/gen/densityfunction/DensityFunction$Noise;sample(DDD)D"))
     private void modifyNoiseSampleArgs(Args args, DensityFunction.NoisePos pos) {
 
-        double d = (pos.blockX() * WorldGenOptions.globalScale[0] + WorldGenOptions.globalOffset[0]) * this.getXzScale() + this.getShiftX().sample(pos);
-        double e = (pos.blockY() * WorldGenOptions.globalScale[1] + WorldGenOptions.globalOffset[1]) * this.getYScale() + this.getShiftY().sample(pos);
-        double f = (pos.blockZ() * WorldGenOptions.globalScale[2] + WorldGenOptions.globalOffset[2]) * this.getXzScale() + this.getShiftZ().sample(pos);
+        double d = (pos.blockX() * UltimateScalerOptions.globalScale[0] + UltimateScalerOptions.globalOffset[0]) * this.getXzScale() + this.getShiftX().sample(pos);
+        double e = (pos.blockY() * UltimateScalerOptions.globalScale[1] + UltimateScalerOptions.globalOffset[1]) * this.getYScale() + this.getShiftY().sample(pos);
+        double f = (pos.blockZ() * UltimateScalerOptions.globalScale[2] + UltimateScalerOptions.globalOffset[2]) * this.getXzScale() + this.getShiftZ().sample(pos);
 
         args.set(0, d);
         args.set(1, e);